requirements.txt
# Web
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
pydantic-settings==2.3.0

# Ingestion
arxiv==2.1.0
nltk==3.8.1

# Embeddings
sentence-transformers==3.0.0
openai==1.35.0

# Storage
psycopg2-binary==2.9.9
pgvector==0.3.1
sqlalchemy==2.0.31
elasticsearch==8.14.0

# LLM
anthropic==0.29.0
langchain==0.2.6
langchain-community==0.2.6

# Eval (Phase 4 — install now, use later)
ragas==0.1.14
mlflow==2.14.0

# Dev
pytest==8.2.2
pytest-asyncio==0.23.7
httpx==0.27.0
structlog==24.2.0

asyncpg==0.29.0
